requirements.txt
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
crispy-bootstrap5==0.7
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==4.2.6
django-countries==7.5.1
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-ranged-response==0.2.0
django-redis==5.4.0
django-simple-captcha==0.5.20
django-taggit==4.0.0
oauthlib==3.2.2
Pillow==10.1.0
psycopg2-binary==2.9.9
redis==5.0.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
gunicorn